Added: - Login + UI - Register (customer + dealership) + UI - 404 page - Security scopes for roles - Dashboard for both customers + dealerships with menus - Redirect to different route based on role after successfully authentication

+++ src/main/java/com/example/car_dealership/repository/DealershipRepository.java
@@ -3,6 +3,5 @@ package com.example.car_dealership.repository;
 import com.example.car_dealership.model.DealerShip;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-
-public interface DealershipRepository extends JpaRepository<DealerShip,Long> {
+public interface DealershipRepository extends JpaRepository<DealerShip, Integer> {
 }
